# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: Thoughtlog REST api documentation
# during dev, should point to your local machine
#host: ec2-107-23-234-81.compute-1.amazonaws.com
host: localhost:10010
# basePath prefixes all resource paths
#basePath: /api
basePath: /
# tagss
tags:
- name: User
  description: "User Authentication Management"
- name: Thoughts
  description: "Thoughts related endpoints"
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
  - multipart/form-data
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: Every request should include api_key in header.
paths:
  /users/signup:
    x-swagger-router-controller: users_controller
    post:
      tags:
        - User
      summary: Allows a user to signup to Thoughtlog
      operationId: signup
      # movie info to be stored
      parameters:
        - in: body
          name: body
          description: User Signup
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserCreateResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorsSignin"
  /users/profile:
    x-swagger-router-controller: users_controller
    get:
      security:
        - Bearer: []
      tags:
        - User
      summary: API for fetching a users profile
      operationId: profile
      parameters:
        - $ref: '#/parameters/AuthorizationHeader'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserProfileResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/signin:
    x-swagger-router-controller: users_controller
    post:
      tags:
        - User
      summary: Allows a user to signin to Thoughtlog
      operationId: signin
      parameters:
        - in: body
          name: body
          description: User Signin
          required: true
          schema:
            $ref: "#/definitions/UserSignin"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserProfileResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorsSignin"
  /users/logout:
    x-swagger-router-controller: users_controller
    delete:
      security:
        - Bearer: []
      tags:
        - User
      summary: Logsout a user on a specific device
      operationId: logout
      parameters:
        - $ref: '#/parameters/AuthorizationHeader'
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /thoughts:
    x-swagger-router-controller: thoughts_controller
    get:
      security:
        - Bearer: []
      tags:
        - Thoughts
      summary: API for fetching all the thoughts
      operationId: fetchAllThoughts
      parameters:
        - $ref: "#/parameters/Page"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/Tags"
        - $ref: "#/parameters/UserId"
        - $ref: "#/parameters/Status"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Thoughts"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      security:
        - Bearer: []
      tags:
      - Thoughts
      summary: API for creating a thought
      operationId: createThought
      parameters:
        - in: formData
          name: description
          type: string
          required: false
          description: Description of the thought.
        - in: formData
          name: category
          type: string
          required: false
          description: Category
        - in: formData
          name: status
          type: string
          required: false
          description: Status
        - in: formData
          name: tags
          type: array
          items:
            type: string
          required: false
          description: Tags
        - in: formData
          name: attachment1
          type: file
          required: false
        - in: formData
          name: attachment2
          type: file
          required: false
        - in: formData
          name: attachment3
          type: file
          required: false
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Thought"
  /thoughts/new:
    x-swagger-router-controller: thoughts_controller
    get:
      security:
        - Bearer: []
      tags:
      - Thoughts
      summary: API for fetching a single thought
      operationId: newThought
      responses:
        "200":
            description: Success
            schema:
              $ref: "#/definitions/NewThought"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /thought/{id}:
    x-swagger-router-controller: thoughts_controller
    get:
      security:
        - Bearer: []
      tags:
      - Thoughts
      summary: API for fetching a single thought
      operationId: showThought
      parameters:
        - in: path
          name: id
          type: string
          description: Thought Id
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Thought"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      security:
        - Bearer: []
      tags:
        - Thoughts
      summary: For removing a thought
      operationId: removeThought
      parameters:
        - in: path
          name: id
          type: string
          description: Thought Id
          required: true
      responses:
        "200":
            description: Success
            schema:
              $ref: "#/definitions/Thought"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      security:
        - Bearer: []
      tags:
        - Thoughts
      summary: For updating a thought.
      operationId: updateThought
      parameters:
        - in: path
          name: id
          type: string
          description: Thought Id
          required: true
        - in: body
          name: thought
          description: Parameters needed for thought updation
          schema:
            $ref: "#/definitions/ThoughtCreateParams"
      responses:
        "200":
            description: Success
            schema:
              $ref: "#/definitions/Thought"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
parameters:
  AuthorizationHeader:
    name: Authorization
    type: string
    in: header
    required: false
  Page:
    name: page
    description: Page number
    type: integer
    in: query
    default: 1
  Limit:
    name: limit
    description: No of records we want while fetch
    type: integer
    in: query
    default: 10
  Tags:
    name: tags
    description: Query tags
    type: string
    in: query
  UserId:
    name: user_id
    description: Id of the user
    type: string
    in: query
  Status:
    name: status
    description: status of the thought
    in: query
    default: 'published'
    enum: ['published', 'drafted', 'deleted']
    type: string
definitions:
  ErrorResponse:
    type: object
    required:
      - errors
    properties:
      status:
        type: string
      errors:
        type: array
        items:
          type: string
  UserCreateResponse:
    type: object
    required:
        - message
        - token
    properties:
      message:
        type: string
      token:
        type: string
  UserProfileResponse:
    type: object
    properties:
      _id:
        type: string
      email:
        type: string
      username:
        type: string
  User:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
      - email
      - password
      - username
  UserSignin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
  ThoughtCreateParams:
    properties:
      description:
        type: string
      status:
        type: string
        enum: ['published', 'deleted', 'drafted']
      category:
        type: string
        enum: ['self', 'bookmark']
      tags:
        type: array
        items:
          type: string
      attachments:
        type: array
        items:
          type: string
      score:
        type: number
    required:
      - description
      - status
      - category
  Thought:
    type:
      object
    properties:
      _id: 
        type: string
      description:
        type: string
      category:
        type: string
      status:
        type: string
      tags:
        type: array
        items:
          type: string
      attachments:
        type: array
        items:
          type: string
      score:
        type: string
      updated_at:
        type: string
    required:
      - _id
      - description
      - category
      - status
      - tags
      - score
      - updated_at
  Thoughts:
    type: object
    properties:
      thoughts:
        type: array
        items:
          $ref: "#/definitions/Thought"
      all_thoughts:
        type: integer
      total_pages:
        type: integer
      current_page:
        type: integer
      limit:
        type: integer
    required:
      - thoughts
      - all_thoughts
      - total_pages
      - current_page
      - limit
  ErrorsSignin:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
  NewThought:
    type: object
    properties:
      id:
        type: string
 